Fix wallet provider chain switching
Fixes an issue in the WalletProvider component where the chain switching logic was not working correctly. Now, when the network option is set to 'testnet', the component will switch to the Shasta Testnet chain if the current chain ID is not '0x3e9'. If the network option is set to 'mainnet', the component will switch to the Tron Mainnet chain if the current chain ID is not '0x2b6653dc'. Additionally, an error toast will be displayed if the current chain ID is '0xcd8690dc' when the network option is set to 'testnet'.